requirements.txt
requests
tavily-python
groq
python-dateutil
python-dotenv
langchain
langchain-core
pydantic
pandas